+++ sdg/stats.py
@@ -9,10 +9,9 @@ Aggregate information for reporting statistics
 import pandas as pd
 
 
-def reporting_status(reporting_status_types, all_meta, extra_fields=None):
+def reporting_status(all_meta, extra_fields=None):
     """
     Args:
-        reporting_status_types: A list of value/label status types
         all_meta: A dictionary containing all metadata items
         extra_fields: List of fields to group stats by, in addition to goal
 
@@ -27,11 +26,12 @@ def reporting_status(reporting_status_types, all_meta, extra_fields=None):
         grouping_fields.append('sdg_goal')
 
     # Generate a report of the possible statuses.
-    status_values = [status['value'] for status in reporting_status_types]
-    value_translation = {status['value']: status['label'] for status in reporting_status_types}
-    status_report = [{'value': status,
-                      'translation_key': value_translation[status]}
-                    for status in status_values]
+    status_values_by_type = {}
+    for indicator_id in all_meta:
+        if 'reporting_status' in all_meta[indicator_id]:
+            status_values_by_type[all_meta[indicator_id]['reporting_status']] = True
+    status_values = list(status_values_by_type.keys())
+    status_report = [{'value': status } for status in status_values]
 
     # Omit any standalone indicators.
     indicators = {k: v for (k, v) in all_meta.items() if 'standalone' not in v or v['standalone'] == False }
@@ -71,7 +71,6 @@ def reporting_status(reporting_status_types, all_meta, extra_fields=None):
     def one_status_report(g, status):
         count = g.get(status, 0)  # If status is missing use 0
         return {'status': status,
-                'translation_key': value_translation[status],
                 'count': count,
                 'percentage': round(100 * count / g['total'],3)}